rename checkpoint dir to store
