+++ go.mod
@@ -7,6 +7,7 @@ require (
 	github.com/go-sql-driver/mysql v1.4.1
 	github.com/lib/pq v0.0.0-20180523175426-90697d60dd84
 	github.com/pkg/errors v0.8.0
+	github.com/stretchr/testify v1.3.0 // indirect
 	google.golang.org/appengine v1.6.1 // indirect
 	gopkg.in/DATA-DOG/go-sqlmock.v1 v1.3.0
 	gopkg.in/redis.v5 v5.2.9
