+++ store/ddb/ddb.go
